% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_calc.R
\name{my_calc}
\alias{my_calc}
\title{my calculator function}
\usage{
my_calc(x, y, fun)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}

\item{fun}{operation to be performed on the vectors}
}
\description{
This function allows you to do basic calculations like sum, subtract,
divide and multiple on vector data
}
