% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonize_labels.R
\name{polygonize_labels}
\alias{polygonize_labels}
\title{Polygonize labelled predictions}
\usage{
polygonize_labels(segment_path, shrink_buffer, expand_buffer, keep)
}
\arguments{
\item{segment_path}{Path to geotiff containing labels}

\item{shrink_buffer}{Negative buffer to reduce polygons by}

\item{expand_buffer}{Postive buffer to grow polygons by after shrinking}

\item{keep}{A proportion of vertices to keep when simplifying polygons}
}
\description{
Takes labelled predictions and converts them to polygons,
breaking them up and simplifying them in the process
}
\details{
This function assumes a GCS
}
